remove hook impl from GitHubStreamComponent
