+++ conf/application.conf
@@ -631,4 +631,7 @@ play.modules.enabled += "com.commercetools.sunrise.common.pages.FrameworkPageMod
 play.modules.enabled += "didemo.DiDemoModule"
 play.http.errorHandler = "com.commercetools.sunrise.http.SunriseDefaultHttpErrorHandler"
 
+GitHubStreamComponent.url = "https://api.github.com/repos/commercetools/commercetools-jvm-sdk/issues"
+GitHubStreamComponent.templateName = "components/GitHub/issues"
+
 include "dev"
